refactor(detection): make geometry dispatch module private
`geometry_dispatch.py` sat at a top-level, non-underscored path
under `detection/`, making it directly importable
(`supervision.detection.geometry_dispatch`) despite the PR's stated
intent that these dispatch helpers stay module-private with no
public API surface. Renamed to `_geometry_dispatch.py`, the standard
convention for signalling "internal, not part of the public API,"
and updated the two import sites.

[resolve group] PR #2374 — item #6 (@Copilot,
https://github.com/roboflow/supervision/pull/2374#discussion_r3618111920)
