+++ src/supervision/detection/_geometry_dispatch.py
